refactor and add active_decorator
