+++ Gemfile
@@ -42,6 +42,8 @@ gem 'rmagick'
 
 gem 'handlebars_assets'
 
+gem 'active_decorator'
+
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
