+++ app/controllers/answer_controller.rb
@@ -2,35 +2,19 @@ class AnswerController < ApplicationController
   def create
     s = Surveys.current_group(request.subdomain).eager_load(:questions, :collaborators).find_by({
       id: params[:survey_id],
-      questions: {
-        id: params[:question_id]
-      },
-      collaborators: {
-        uuid: session[:uuid]
-      }
+      questions:     {id:   params[:question_id]},
+      collaborators: {uuid: session[:uuid]}
     })
     q = s.questions.first
     c = s.collaborators.first
 
     if params[:answer].present?
-      if q.date?
-        t = {}
-        q.each_date.each {|d| t[d] = params[d]}
-        params[:answer] = {
-          text: {
-            name:    params[:name],
-            comment: params[:comment],
-            date:    t,
-          }.to_json
-        }
-      end
-
       a = c.answers.where(question_id: q.id).first_or_initialize answer_params
-      # override text if question is multiple
-      a.text = params[:answer][:text].to_json if q.multiple?
-      a.text = params[:answer][:text] if q.date?
+      # override text if question is multiple or date
+      a.text = params[:answer][:text].to_json if q.multiple? || q.date?
       a.save!
     else
+      # if no answer pathed destroy all answers for this question.
       c.answers.where(question_id: q.id).destroy_all
     end
 
@@ -42,7 +26,6 @@ class AnswerController < ApplicationController
   alias_method :update, :create
 
   private
-
   def answer_params
     params.require(:answer).permit :text
   end
