+++ app/views/question/_date.html.slim
@@ -2,7 +2,7 @@
 h2.ui.header
   | 表示名
 .field
-  = text_field_tag :name, ans['name']
+  = text_field_tag 'answer[text][name]', ans['name']
 h2.ui.header
   | 日にち候補
 - @question.each_date.each.with_index(1) do |v, i|
@@ -19,8 +19,8 @@ h2.ui.header
               | △
             button type="button" class="ui huge toggle basic button#{d[v] == '0' ? ' active' : ''}" data-val="0"
               | ☓
-            input type="hidden" name="#{v}" value="#{d[v]}"
+            input type="hidden" name="answer[text][date][#{v}]" value="#{d[v]}"
 h2.ui.header
   | コメント
 .field
-  input type="text" name="comment" value="#{ans['comment']}"
+  = text_field_tag 'answer[text][comment]', ans['comment']
