+++ spec/decorators/question_decorator_spec.rb
@@ -1,0 +1,8 @@
+# coding: utf-8
+require 'rails_helper'
+
+describe QuestionDecorator do
+  let(:question) { Question.new.extend QuestionDecorator }
+  subject { question }
+  it { should be_a Question }
+end
